lint

+++ src/Calculator.sol
@@ -8,27 +8,27 @@ contract Calculator {
         _number = number_;
     }
 
-    function add(uint x) public returns(uint256) {
+    function add(uint256 x) public returns (uint256) {
         _number += x;
         return _number;
     }
 
-    function sub(uint x) public returns(uint256) {
+    function sub(uint256 x) public returns (uint256) {
         _number -= x;
         return _number;
     }
 
-    function mult(uint x) public returns(uint256) {
+    function mult(uint256 x) public returns (uint256) {
         _number *= x;
         return _number;
     }
 
-    function div(uint x) public returns(uint256) {
+    function div(uint256 x) public returns (uint256) {
         _number /= x;
         return _number;
     }
 
-    function getNumber() public view returns(uint256) {
+    function getNumber() public view returns (uint256) {
         return _number;
     }
 }
